Added logging of prediction results

+++ main.py
@@ -11,6 +11,7 @@ import numpy as np
 import pandas as pd
 from openea.models.basic_model import BasicModel
 
+from dataset.csv_dataset import CsvDataset, CsvType
 from dataset.dataset import Dataset
 from matching.matcher import MatchModelTrainer
 from matching.pair_to_vec import (
@@ -68,19 +69,33 @@ def run_for_dataset(dataset_idx):
         + dataset.labelled_test_pairs
     )
 
-    all_sims, min_max, scale_cols = create_feature_similarity_frame(
-        embeddings, all_pairs, kgs, only_training=True,
-    )
-    output_folder = existing_folder or embedding_model.out_folder[:-1]
-    if not path.exists(output_folder):
-        dir_path = output_folder.split("/")
-        for i in range(len(dir_path)):
-            if not path.exists("/".join(dir_path[: i + 1])):
-                os.mkdir("/".join(dir_path[: i + 1]))
-    all_sims.to_parquet(f"{output_folder}/all_sims.parquet")
-    joblib.dump(min_max, f"{output_folder}/min_max.pkl")
-    with open(f"{output_folder}/scale_cols.json", "w") as f:
-        json.dump(scale_cols, f)
+    if (
+        False and
+        existing_folder is not None
+        and path.exists(f"{existing_folder}/all_sims.parquet")
+        and path.exists(f"{existing_folder}/min_max.pkl")
+        and path.exists(f"{existing_folder}/scale_cols.json")
+    ):
+        print("found existing similarity frame data")
+        all_sims = pd.read_parquet(f"{existing_folder}/all_sims.parquet")
+        min_max = joblib.load(f"{existing_folder}/min_max.pkl")
+        with open(f"{existing_folder}/scale_cols.json") as f:
+            scale_cols = json.load(f)
+    else:
+        all_sims, min_max, scale_cols = create_feature_similarity_frame(
+            embeddings, all_pairs, kgs, only_training=True,
+        )
+        output_folder = existing_folder or embedding_model.out_folder[:-1]
+        if not path.exists(output_folder):
+            dir_path = output_folder.split("/")
+            for i in range(len(dir_path)):
+                if not path.exists("/".join(dir_path[: i + 1])):
+                    os.mkdir("/".join(dir_path[: i + 1]))
+        all_sims.to_parquet(f"{output_folder}/all_sims.parquet")
+        joblib.dump(min_max, f"{output_folder}/min_max.pkl")
+        with open(f"{output_folder}/scale_cols.json", "w") as f:
+            json.dump(scale_cols, f)
+
     all_sims = all_sims.dropna(axis=1, how="all", thresh=int(0.1 * len(all_sims)))
 
     pair_to_vecs = [
@@ -91,9 +106,11 @@ def run_for_dataset(dataset_idx):
         OnlyEmb(embeddings, all_sims, min_max, scale_cols, kgs),
     ]
 
-    run_params = [(model_fac(pair_to_vec), dataset, pair_to_vec)
-                  for pair_to_vec in pair_to_vecs
-                  for model_fac in model_factories]
+    run_params = [
+        (model_fac(pair_to_vec), dataset, pair_to_vec)
+        for pair_to_vec in pair_to_vecs
+        for model_fac in model_factories
+    ]
 
     with Pool(processes=4) as pool:
         results_list = pool.starmap(run, run_params)
@@ -117,18 +134,35 @@ def run_for_dataset(dataset_idx):
         ],
     )
 
+    test_pred = [
+        ("test", r["model_name"], r["vector_name"], r["test_prediction"])
+        for r in results_list
+    ]
+    train_pred = [
+        ("train", r["model_name"], r["vector_name"], r["train_prediction"])
+        for r in results_list
+    ]
+    val_pred = [
+        ("val", r["model_name"], r["vector_name"], r["val_prediction"])
+        for r in results_list
+    ]
+    for typ, model, vec, pred in test_pred + train_pred + val_pred:
+        pd.DataFrame(data=pred, columns=["left", "right", "val", "pred"]).to_csv(
+            csv_result_file.replace(".csv", f"_{model}_{vec}_{typ}_pred.csv"),
+            index=False,
+        )
+
     if path.exists(csv_result_file):
         old_frame = pd.read_csv(csv_result_file)
         results = merge_dataframes(old_frame, results)
     results.to_csv(csv_result_file, index=False)
 
 
-def merge_dataframes(old_frame, new_frame):
+def merge_dataframes(old_frame: pd.DataFrame, new_frame):
     old_frame.set_index(keys=["model_name", "vector_name"], drop=False, inplace=True)
     new_frame.set_index(keys=["model_name", "vector_name"], drop=False, inplace=True)
 
-    old_frame.update(new_frame)
-    return old_frame
+    return new_frame.combine_first(old_frame)
 
 
 def find_existing_result_folder(embedding_model: BasicModel):
@@ -189,12 +223,15 @@ def run(
         "train_precision": train_eval.precision,
         "train_recall": train_eval.recall,
         "train_f1": train_eval.f1,
+        "train_prediction": train_eval.prediction,
         "val_precision": valid_eval.precision,
         "val_recall": valid_eval.recall,
         "val_f1": valid_eval.f1,
+        "val_prediction": valid_eval.prediction,
         "test_precision": test_eval.precision,
         "test_recall": test_eval.recall,
         "test_f1": test_eval.f1,
+        "test_prediction": test_eval.prediction,
         "train_time": train_time,
         "test_time": test_time,
     }
